+++ app/config/global.cpp
@@ -96,4 +96,17 @@ namespace Config {
   bool Global::singleInstance() const {
     return storage.get("single_instance").get<bool>();
   }
+
+  int Global::playbackLogSize() const {
+    return storage.get("playback_log_size").get<int>();
+  }
+
+  PlaylistUi::ColumnsConfig Global::columnsConfig() const {
+    auto raw = storage.get("columns_config").get<QList<Config::Value>>();
+    return PlaylistUi::ColumnsConfig::deserialize(raw);
+  }
+
+  bool Global::saveColumnsConfig(const PlaylistUi::ColumnsConfig &arg) {
+    return storage.set("columns_config", arg.serialize());
+  }
 }