+++ app/config/global.h
@@ -3,6 +3,7 @@
 
 #include "storage.h"
 #include "sort_ui/sortingpresetsdialog.h"
+#include "playlist_ui/columnsconfig.h"
 
 #include <QPair>
 #include <QList>
@@ -38,6 +39,11 @@ namespace Config {
     bool saveIpcPort(int arg);
 
     bool singleInstance() const;
+
+    int playbackLogSize() const;
+
+    PlaylistUi::ColumnsConfig columnsConfig() const;
+    bool saveColumnsConfig(const PlaylistUi::ColumnsConfig &arg);
 
   private:
     Config::Storage storage;