+++ app/shortcuts.cpp
@@ -1,6 +1,6 @@
 #include "shortcuts.h"
 
-Shortcuts::Shortcuts(QWidget *parent, Config::Global &global_c) : QObject(parent), global_conf(global_c),
+Shortcuts::Shortcuts(QWidget *parent) : QObject(parent),
   _quit(parent),
   _focus_library(parent),
   _focus_playlists(parent),
@@ -16,6 +16,7 @@ Shortcuts::Shortcuts(QWidget *parent, Config::Global &global_c) : QObject(parent
   _open_main_menu(parent),
   _open_playback_log(parent),
   _open_sort_menu(parent),
+  _open_shortcuts_menu(parent),
   _play_global(parent),
   _pause_global(parent),
   _stop_global(parent),
@@ -24,6 +25,27 @@ Shortcuts::Shortcuts(QWidget *parent, Config::Global &global_c) : QObject(parent
 {
   setupLocal();
   setupGlobal();
+}
+
+QVector<QPair<QString, QString> > Shortcuts::describe() const {
+  QVector<QPair<QString, QString> > r;
+  r << QPair<QString, QString>(tr("Play"), _play.key().toString());
+  r << QPair<QString, QString>(tr("Stop"), _stop.key().toString());
+  r << QPair<QString, QString>(tr("Pause"), _pause.key().toString());
+  r << QPair<QString, QString>(tr("Next"), _next.key().toString());
+  r << QPair<QString, QString>(tr("Previous"), _prev.key().toString());
+  r << QPair<QString, QString>(tr("Focus on library"), _focus_library.key().toString());
+  r << QPair<QString, QString>(tr("Focus on playlists"), _focus_playlists.key().toString());
+  r << QPair<QString, QString>(tr("Focus on playlist"), _focus_playlist.key().toString());
+  r << QPair<QString, QString>(tr("Focus on library filter"), _focus_filter_library.key().toString());
+  r << QPair<QString, QString>(tr("Focus on playlists filter"), _focus_filter_playlists.key().toString());
+  r << QPair<QString, QString>(tr("Focus on playlist filter"), _focus_filter_playlist.key().toString());
+  r << QPair<QString, QString>(tr("Open main menu"), _open_main_menu.key().toString());
+  r << QPair<QString, QString>(tr("Open playback log"), _open_playback_log.key().toString());
+  r << QPair<QString, QString>(tr("Open sort menu"), _open_sort_menu.key().toString());
+  r << QPair<QString, QString>(tr("Open shortcuts dialog"), _open_shortcuts_menu.key().toString());
+  r << QPair<QString, QString>(tr("Quit"), _quit.key().toString());
+  return r;
 }
 
 void Shortcuts::setupGlobal() {
@@ -56,6 +78,7 @@ void Shortcuts::setupLocal() {
   connect(&_open_main_menu, &QShortcut::activated, this, &Shortcuts::openMainMenu);
   connect(&_open_playback_log, &QShortcut::activated, this, &Shortcuts::openPlabackLog);
   connect(&_open_sort_menu, &QShortcut::activated, this, &Shortcuts::openSortMenu);
+  connect(&_open_shortcuts_menu, &QShortcut::activated, this, &Shortcuts::openShortcutsMenu);
 
   _quit.setKey(Qt::CTRL | Qt::Key_Q);
   _focus_library.setKey(Qt::CTRL | Qt::Key_1);
@@ -72,4 +95,5 @@ void Shortcuts::setupLocal() {
   _open_main_menu.setKey(Qt::ALT | Qt::Key_M);
   _open_playback_log.setKey(Qt::CTRL | Qt::Key_L);
   _open_sort_menu.setKey(Qt::CTRL | Qt::Key_S);
+  _open_shortcuts_menu.setKey(Qt::ALT | Qt::Key_S);
 }